TES5: Add Import Factions support
No idea why this was using any of that nonsense, it's just a dead simple
'preserves' patcher and can therefore use the 100% standard
_SimpleImporter pattern. It looks like it may have been based off of the
Import Relations patcher, and was intended to merge additions and
deletions, but during development its author decided that was a bad
idea and made it a standard preserver. In fact, that seems to be what
the CBash patcher does - however, IMO the 'preserves' behavior makes
more sense here, so we should adjust the CBash patcher at some (distant)
point.

Please excuse the fact that this makes the Oblivion links no longer
sorted, will fix itself once everything is refactored :)