Split cint into multiple modules
Everything was spit according to structure - who needed to import what
and where - and not function. The latter would need a more complete
rework of this package which, depending on what is decided on CBash #27
may be unnecessary.

The falloutnv module is non-functional and is therefore never imported -
less 8k lines to import. To do this, ObCollection needed to lose its
FnvModFile reference, which was replaced with an assertion that it is an
ObModFile. Not sure if we rely on this check somewhere, but if we do,
we shouldn't.

Aside from the above, no other changes were made to any variable,
function or class besides splitting.