New feature: Lock Active Plugins RRR
Just an enhancement for Lock Load Order. If both are toggled on, changes
to the enabled/disabled state of plugins will be undone on boot/refresh
as well.

Enabled it by default, since people who have Lock Load Order on are most
likely pure WB users who will also want this feature on (and may even be
assuming that WB does this already).

Under # 210 <--- RRR