bolt.RecPath
Maybe brec.RecPath? But brec is already so big, plus I want to use it in
tweaks (for now, eventually everywhere ;)) and bolt is a nice place for
such things...

Anyway, this is a general way to get/set record attributes, with
validation and pre-processing to speed things up. Will write a wiki page
for it (yes yes I know, will get to the record classes too).