Implement Tweak Pooling
Tested with all Oblivion tweaks, builds fine but had a few hundred
records regressed vs nightly when I last checked (out of 11000+, so not
bad).

Note just how similar each of the buildPatch overrides in the tweaks
look now - not an accident, will be handled in the rest of this branch ;)

No assorted tweaks use DynamicNamedTweak anymore. It became an MRO
nightmare in there, so I binned them. The new static solution does take
more LOC, but I think it's worth it.

The GMST tweakers *really* don't conform nicely to this new
implementation, so that's the only file that increased in line count.

Note weird setdefault in 'merge similar eyes' patchers, only difference
to 'merge similar hairs' algorithm, but shouldn't have done anything in
that place anyway?

I unified a lot of logs along the way, the result seems to look a lot
better :)
Quite a few bugs no one knew existed got fixed as well, e.g. CBash
always setting static attenuation to 0, regardless of what the user
chose, because of an integer division by 100.

One wart is that a few tweaks now duplicate work in wants_record and
buildPatch/apply. Will be tackled later in this branch.

Fog fix: do proper float comparison

Fixes a few edge cases where it doesn't do the fix even though it
should, because a plugin contains a value really close but not exactly
equal to 0.0. Obviously can't reproduce this, my graphics card is from
Nvidia but I get no 'fog problem' with it. Was reported on Discord, and
this does fix it for the missed records that were reported, so...

fixup!     Implement Tweak Pooling

Rename rec_sig will be used for recType - 0 occurrences remain SSS

Move _any_body_flag_set in _APlayableTweak