Move MreHasEffects to Oblivion
No idea why this got added to the other games, the 'effects' in Oblivion
are completely different from the ones in the later games. Most notably,
in Oblivion it is not possible to create new MGEFs and they are uniquely
identified by their 4-letter EDID.

Anyways, this class is fundamentally Oblivion-only and makes no sense
for the later games, so keep it in game/oblivion. Was only used in
Oblivion-specific patchers and save editing code.