Split brec to a package
Names could use some work, but editing is much nicer. Did it in a single
commit so I wouldn't have to bother with fixing imports several times
over.

Only had to fix one circular dependency (arose due to the way I designed
the distributor API, could be fixed by changing it - but I like the way
it works :)).

Added myself as author of advanced_elements, since that file is 100% my
code - the rest is a mix of my code and code that's been around for a
long time.