XXX installer settings version 2:
- stop pickling instances of the Installer subclasses - pickle a tuple
of persistent attributes and recreate the object. Those classes are rather
unflexible and we need to extend BAIN at some point. bleeding edge - I also
considered adding kwargs (but this persists the names of the attributes that
even this implementation avoids) or maybe use a Translator_ to move the
classes from bosh - but this is not the real problem, the real problem is
Installer internal API and the goal is to simplify it so we don't at least
have to reason about how we are pickling it