bain: Offer to redirect script sources in SSE
In SSE Bethesda made the mistake of packaging the CK's script source
as 'source/scripts' instead of 'scripts/source'. The CK even still
expects the sources to be in 'scripts/source', so you'd have to edit
its INI if you wanted to use 'source/scripts'. However, some modders
have nonetheless adopted this convention, so to support this we
redirect them to the correct path while scanning the package in BAIN.

Also threw in some much needed refactoring for the weird skip links
logic and split the remapping/renaming logic out of refreshDataSizeCrc
(very welcome, that method is big enough as is).