tests: Add initial test files
Deleted all but test_bolt for now - were either completely outdated
or wouldn't ever work on CI as is, which is the primary goal of this
experiment.

Also added actual tests for OrderedLowerDict. Had to move to pytest
because unittest just does *not* want to work with our folder structure.
pytest, on the other hand, works right out of the box :)
(I also like pytest a lot more than unittest, so I might be biased)